Internal tidy
- Change re-export paths. Make sure no public type is re-exported twice.
- Split tensor/mod.rs into small modules.
