+++ src/lib.rs
@@ -11,19 +11,19 @@ mod error;
 pub use error::TchError;
 
 pub(crate) mod wrappers;
-pub use wrappers::device::{Cuda, Device};
-pub use wrappers::jit::{CModule, IValue};
-pub use wrappers::kind::Kind;
-pub use wrappers::scalar::Scalar;
 pub use wrappers::{
-    get_num_interop_threads, get_num_threads, manual_seed, set_num_interop_threads, set_num_threads,
+    device::{Cuda, Device},
+    get_num_interop_threads, get_num_threads,
+    jit::{CModule, IValue},
+    kind::Kind,
+    manual_seed,
+    scalar::Scalar,
+    set_num_interop_threads, set_num_threads,
+    tensor::{no_grad, no_grad_guard, NoGradGuard, Reduction, Tensor},
 };
 
 mod tensor;
-pub use tensor::{
-    index, no_grad, no_grad_guard, IndexOp, NewAxis, NoGradGuard, Reduction, Shape, Tensor,
-    TensorIndexer,
-};
+pub use tensor::{index, IndexOp, NewAxis, Shape, TensorIndexer};
 
 pub mod nn;
 pub mod vision;
